requirements.txt
Flask
numpy
Pillow
tensorflow
opencv-python